give design attributes: interventional and observational model - the same naming structure.  Force a decision:  do we use the attribute names provided the clinicaltrials.gov, or, when we find inconsistencies like this, do we correct them?  I vote for correcting them in an effort to make this system as easy and intuitive to use as possible.
Also, remove the attributes that provide dates in string formats.  These fields were initially included because that's how the old AACT provides most date values,so tried to honor previous decisions made to convert dates to strings, but finding these fields to be of little value since we can always convert date-type fields to strings if we want to.